+++ src/main/java/org/beilstein/chemxtract/cdx/datatypes/CDConstraintType.java
@@ -28,5 +28,5 @@ public enum CDConstraintType {
 
   Distance,
   Angle,
-  ExclusionSphere;
+  ExclusionSphere
 }