+++ src/main/java/org/beilstein/chemxtract/cdx/datatypes/CDLabelDisplay.java
@@ -36,5 +36,5 @@ public enum CDLabelDisplay {
   /** Label is stacked below */
   Below,
   /** Best initial alignment */
-  BestInitial;
+  BestInitial
 }