+++ src/main/java/org/beilstein/chemxtract/cdx/datatypes/CDPositioningType.java
@@ -26,5 +26,5 @@ public enum CDPositioningType {
   Auto,
   Angle,
   Offset,
-  Absolute;
+  Absolute
 }