+++ src/main/java/org/beilstein/chemxtract/cdx/datatypes/CDSymbolType.java
@@ -50,5 +50,5 @@ public enum CDSymbolType {
   /** Relative */
   Relative,
   /** TODO: find documentation */
-  LonePairBar;
+  LonePairBar
 }