+++ src/main/java/org/beilstein/chemxtract/samples/BCXTractSubstances.java
@@ -44,7 +44,7 @@ import org.openscience.cdk.smiles.SmilesParser;
  */
 public class BCXTractSubstances {
 
-  public static void main(String[] args) throws Exception {
+  void main(String[] args) throws Exception {
     if (args.length != 1) {
       System.err.println("Input CDX file must be given as argument.");
       System.exit(1);
@@ -91,6 +91,6 @@ public class BCXTractSubstances {
       i++;
     }
 
-    System.out.println("\n\n" + i + " substances extracted to current working directory.");
+    IO.println("\n\n" + i + " substances extracted to current working directory.");
   }
 }